requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
pip==24.0
setuptools==65.5.0
sqlparse==0.5.3
tzdata==2025.2
image==1.5.33